+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/main/popup/newControl/NewControlPopup.java
@@ -73,7 +73,7 @@ public class NewControlPopup extends StagePopup<VBox> {
 			}
 		});
 		controller.cobBaseControl.getSelectionModel().select(ControlType.STATIC);
-		myRootElement.getChildren().addAll(new Separator(Orientation.HORIZONTAL), getResponseFooter(true, true, true));
+		myRootElement.getChildren().addAll(new Separator(Orientation.HORIZONTAL), getBoundResponseFooter(true, true, true));
 		myStage.sizeToScene();
 	}